devide project into two pages: past-projects, active-projects
